requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
Django==3.2.8
django-elasticsearch-dsl==7.2.1
django-js-asset==1.2.2
django-mptt==0.13.4
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
factory-boy==3.2.0
Faker==10.0.0
h11==0.12.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
mypy-extensions==0.4.3
outcome==1.1.0
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
wsproto==1.0.0
djangorestframework~=3.13.1